+++ src/Transformer/TextMatchesTransformer.php
@@ -29,7 +29,7 @@ class TextMatchesTransformer extends AbstractTransformer
      */
     protected function transform() : void
     {
-        $keyLocator = array_keys($this->step['textMatches'])[1];
+        $keyLocator = array_keys($this->step['textMatches'])[0];
         $errorMessage = sprintf('%s[%s] text does not match %s', $keyLocator, $this->step['textMatches'][$keyLocator], $this->step['textMatches']['regexp']);
         $this->driver->wait(Configuration::get('wait.for'), Configuration::get('wait.every'))
                      ->until(WebDriverExpectedCondition::elementTextMatches(