+++ src/Transformer/Logic/LogicExecutor.php
@@ -95,7 +95,7 @@ class LogicExecutor
         try {
             $this->answer = $logic->answeredBy();
         } catch (\Exception $e) {
-            if(null !== $this->eventDispatcher) {
+            if (null !== $this->eventDispatcher) {
                 $this->eventDispatcher->notify(AutoMateEvents::LOGIC_EXCEPTION, ['exception' => $e]);
             }
             $this->answer = $this->valueAtException;