+++ tests/unit/Handler/PageHandlerTest.php
@@ -4,9 +4,10 @@ use Automate\Configuration\Configuration;
 use Automate\Handler\PageHandler;
 use PHPUnit\Framework\TestCase;
 
-class PageHandlerTest extends TestCase {
-
-    public function testShouldLoadPageHandler() {
+class PageHandlerTest extends TestCase
+{
+    public function testShouldLoadPageHandler()
+    {
         Configuration::load(__DIR__.'/../files/config-test.yaml');
         
         PageHandler::load('tests.index', __DIR__.'/../files/pages');